requirements.txt
Pillow>=10.4.0
beautifulsoup4>=4.12.3
feedparser>=6.0.11
google-api-python-client>=2.137.0
google-auth-oauthlib>=1.2.1
google-auth>=2.34.0
moviepy>=1.0.3
openai>=1.43.0
pydub>=0.25.1
python-dateutil>=2.9.0
python-dotenv>=1.0.1
requests>=2.32.3
srt>=3.5.3
tqdm>=4.66.4
